pt1_proyecto_sd/generator/requirements.txt
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
Flask==2.3.2